requirements.txt
torch
torchvision
numpy
python-chess
pygame